Fix RecurrentNetworkGradient with batch size > 1
Summary:
Fix RecurrentNetworkGradient with batch size > 1.
The main issue was that we always set the Gradient output to 1, 1, recurrent_size which mismatch the input (1, batch_size, recurrent_size).
Further gradient ops do Squeeze and split assuming that output gradient blob is the same size as the input so they fail.

The fix is simply Resizing the output as the input (1, batch_size, recurrent_size), I had to move the resize to the RunOnDevice since batch_size is computed from Input(0) which is not available till the we actually run the op.

Differential Revision: D4301487

fbshipit-source-id: e5c7426d6e770d985ce72a3737381a2b4af333ba