Normalize rank loss gradient to avoid convergence issues when the number of pairs is really large
Summary:
Essentially, when number of pairs is around 1000, then only positive samples in the list gets a massive boost from all the negative examples. This diff normalizes the gradient and the loss with the number of pairs.

This diff also adds protection against NaN and more logging to help debug.

Reviewed By: kdub0

Differential Revision: D4359782

fbshipit-source-id: 7240344ddb1f2f670d1eec1b03e7f6e413f3dfcc