Fix sparse to dense and improve DispatchHelper
Summary:
Actually adds stuff on duplicated indices. I didn't use UnorderedSegmentSum because it'd need more modifications for figuring out the first dimension and I don't want to make that function more complex than it's already is :)

We theoretically can have a version that does CopyItems and fails on duplicate indices as a fallback. But I haven't implemented it yet as it wouldn't be that useful for now.

Also fixes hypothesis test - doing rand() inside the body is not cool as it makes hypothesis run forever

Differential Revision: D4814574

fbshipit-source-id: 1851ec5f5df8fc4bf4844585076b8af23a06b0b2